requirements.txt
discord.py==1.6.0
boto3==1.17.44
aiorwlock==1.0.0
discord==1.0.1
Pillow==8.2.0
python-dotenv==0.17.0
